+++ test/loaders/args.test.js
@@ -1,4 +1,4 @@
-const assert = require('chai').assert;
+const { ifError, strictEqual: equal } = require('assert');
 const args = require('../../lib/loaders/args');
 const EventEmitter = require('events').EventEmitter;
 
@@ -8,8 +8,8 @@ describe('args', () => {
 
   it('should load arguments', (t, done) => {
     args()(confabulous, (err, config) => {
-      assert.ifError(err);
-      assert.equal(config['test-argument'], true);
+      ifError(err);
+      equal(config['test-argument'], true);
       done();
     });
   });
@@ -21,8 +21,8 @@ describe('args', () => {
         cb(null, config);
       }
     ])(confabulous, (err, config) => {
-      assert.ifError(err);
-      assert.equal(config['test-argument'], false);
+      ifError(err);
+      equal(config['test-argument'], false);
       done();
     });
   });