+++ test/loaders/file.test.js
@@ -1,4 +1,4 @@
-const assert = require('chai').assert;
+const { ok, ifError, strictEqual: equal } = require('assert');
 const file = require('../../lib/loaders/file');
 const fs = require('fs');
 const EventEmitter = require('events').EventEmitter;
@@ -23,44 +23,44 @@ describe('file', () => {
 
   it('should require path when mandatory', (t, done) => {
     file()(confabulous, (err) => {
-      assert(err);
-      assert.equal(err.message, 'path is required');
+      ok(err);
+      equal(err.message, 'path is required');
       done();
     });
   });
 
   it('should load configuration', (t, done) => {
     file({ path: 'test/data/config.json' })(confabulous, (err, text) => {
-      assert.ifError(err);
+      ifError(err);
       const config = JSON.parse(text);
-      assert.equal(config.loaded, 'loaded');
+      equal(config.loaded, 'loaded');
       done();
     });
   });
 
   it('should report missing files when mandatory', (t, done) => {
     file({ path: 'does-not-exist.json' })(confabulous, (err) => {
-      assert(err);
-      assert(/ENOENT/.test(err.message), err.message);
+      ok(err);
+      ok(/ENOENT/.test(err.message), err.message);
       done();
     });
   });
 
   it('should ignore missing files when not mandatory', (t, done) => {
     file({ path: 'does-not-exist.json', mandatory: false })(confabulous, (err) => {
-      assert.equal(err, true);
+      equal(err, true);
       done();
     });
   });
 
   it('should emit change event when content changes', (t, done) => {
     file({ path: 'test/data/config.json', watch: true })(confabulous, (err, text) => {
-      assert.ifError(err);
+      ifError(err);
       const config = JSON.parse(text);
-      assert.equal(config.loaded, 'loaded');
+      equal(config.loaded, 'loaded');
       config.updated = new Date().toISOString();
       fs.writeFile('test/data/config.json', JSON.stringify(config, null, 2), (err) => {
-        assert.ifError(err);
+        ifError(err);
       });
     }).once('change', done);
   });
@@ -68,9 +68,9 @@ describe('file', () => {
   it('should emit change event when file is deleted', (t, done) => {
     fs.writeFileSync(doomed, JSON.stringify({ foo: "bar" }));
     file({ path: doomed, mandatory: false, watch: true })(confabulous, (err) => {
-      assert.ifError(err);
+      ifError(err);
       fs.unlink(doomed, (err) => {
-        assert.ifError(err);
+        ifError(err);
       });
     }).once('change', done);
   });
@@ -82,8 +82,8 @@ describe('file', () => {
         cb(null, JSON.parse(text));
       }
     ])(confabulous, (err, config) => {
-      assert.ifError(err);
-      assert.equal(config.loaded, 'loaded');
+      ifError(err);
+      equal(config.loaded, 'loaded');
       done();
     });
   });