+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - Replace mocha with zUnit
 - Replace lodash.set, lodash.has with dot-prop
 - Replace imperative with esnext style
+- Replace chai with node assert
 
 ## [1.7.0]
 ### Updated